requirements-HF.txt
accelerate
diffusers
invisible_watermark
torch
transformers
xformers
pillow
matplotlib
pathlib